Support LD2451 alongside RD03-D
LD2451 selection previously registered a device but had no firmware
parser or initialization. Complete the protocol path and consolidate the
device configuration, uploader and admin work into one reviewable
change.

Select exactly one firmware module, initialize it at its own baud rate,
validate variable-length LD2451 target frames and command
acknowledgements, and convert polar targets to signed cm/s and 32-bit
millimetres. Keep the RD03-D sentinel behavior specific to RD03-D. Route
configuration and photo uploads by explicit device identity and allow
LD2451 capture settings up to 100 metres. Deployment builds the firmware
feature matching that identity.

Validation: 9 LD2451/shared and 2 RD03-D/shared host tests; 39 uploader
and 68 web tests; both ESP firmware variants build and link. Invalid
feature combinations fail compilation. Deployment argument and shell
checks pass.

Fixtures are synthetic protocol vectors, not physical captures. Real
UART, direction, GPIO42 and camera validation remain required. The
manufacturer contradicts its direction table in an example; follow table
9 and preserve raw direction diagnostics for the hardware check. Deploy
web and uploader together because photo uploads now require device
identity.

+++ uploader/src/fake_radar_reader.rs
@@ -89,8 +89,8 @@ impl Actor for FakeRadarReader {
                 let message = format!(
                     "EVENTS: TARGET: {} {} {}",
                     target.speed_cm_s(),
-                    target.x as i16,
-                    target.y as i16,
+                    target.x as i32,
+                    target.y as i32,
                 );
                 self.radar_input.process_log_message(message);
                 target.step(dt, &mut rng);
